+++ src/reducer.ts
@@ -2,6 +2,7 @@ import { GameQuest, PoiAction, QuestTab } from './poi/types'
 
 const initState = {
   questList: null as null | GameQuest[],
+  observedQuestMap: {} as Record<number, GameQuest>,
   tabId: QuestTab.ALL,
 }
 
@@ -14,10 +15,15 @@ export const reducer = (
   switch (action.type) {
     case '@@Response/kcsapi/api_get_member/questlist': {
       const { body, postBody } = action
+      const observedQuestMap = { ...state.observedQuestMap }
+      body.api_list.forEach((quest) => {
+        observedQuestMap[quest.api_no] = quest
+      })
 
       return {
         ...state,
         questList: body.api_list,
+        observedQuestMap,
         tabId: postBody.api_tab_id,
       }
     }